Fixed issue in generating signed apks

+++ app/src/main/java/com/bry/adcafe/adapters/AdminStatItem.java
@@ -68,7 +68,7 @@ public class AdminStatItem {
         mNumberReached.setText(String.format("Number Reached : %d users.", mAdvert.getNumberOfTimesSeen()));
 
         int numberOfUsersWhoDidntSeeAd = (mAdvert.getNumberOfUsersToReach()- mAdvert.getNumberOfTimesSeen());
-        int ammountToBeRepaid = (int)(numberOfUsersWhoDidntSeeAd*mAdvert.getAmountToPayPerTargetedView());
+        int ammountToBeRepaid = numberOfUsersWhoDidntSeeAd*mAdvert.getAmountToPayPerTargetedView();
         String number = Integer.toString(ammountToBeRepaid);
 
         mAmountToReimburse.setText("Amount to reimburse : "+number+" Ksh.");
@@ -115,7 +115,7 @@ public class AdminStatItem {
                         mAmountToReimburse.setText("Reimbursing amount:  0 Ksh");
                     }else{
                         int numberOfUsersWhoDidntSeeAd = (mAdvert.getNumberOfUsersToReach()- mAdvert.getNumberOfTimesSeen());
-                        int ammountToBeRepaid = (int)(numberOfUsersWhoDidntSeeAd*mAdvert.getAmountToPayPerTargetedView());
+                        int ammountToBeRepaid = numberOfUsersWhoDidntSeeAd*mAdvert.getAmountToPayPerTargetedView();
                         String number = Integer.toString(ammountToBeRepaid);
                         mHasBeenReimbursed.setText("Status: NOT Reimbursed.");
                         mAmountToReimburse.setText("Reimbursing amount: "+number+" Ksh");
